Moved methods from InventoryUtils to MaterialUtils
